requirements.txt
numpy==1.18.2
Pillow==7.1.1
pyo==1.0.1
PyYAML==5.3.1
six==1.14.0
wxPython==4.0.7.post2
